+++ packages/plugin/src/code-gen/service-client-generator-base.ts
@@ -8,15 +8,14 @@ import * as ts from "typescript";
 import * as rpc from "@protobuf-ts/runtime-rpc";
 import {CommentGenerator} from "./comment-generator";
 import {Interpreter} from "../interpreter";
+import {ClientStyle} from "../our-options";
 import assert = require("assert");
 
 
 export abstract class ServiceClientGeneratorBase {
 
 
-    // TODO #8 make style a public property of client? add "style" to reserved name in Interpreter.createTypescriptNameForMethod and add to name-clash-proto
-
-    abstract readonly style: rpc.ClientStyle;
+    abstract readonly style: ClientStyle;
     protected readonly commentGenerator: CommentGenerator;
 
 
@@ -52,7 +51,7 @@ export abstract class ServiceClientGeneratorBase {
     generateInterface(descriptor: ServiceDescriptorProto, source: TypescriptFile): ts.InterfaceDeclaration {
         const
             interpreterType = this.interpreter.getServiceType(descriptor),
-            styleName = rpc.ClientStyle[this.style].toLowerCase(),
+            styleName = ClientStyle[this.style].toLowerCase().replace('_client', ''),
             IServiceClient = this.imports.type(descriptor, `${styleName}-client-interface`);
 
         const methods = interpreterType.methods.map(mi => {
@@ -104,7 +103,7 @@ export abstract class ServiceClientGeneratorBase {
     generateImplementationClass(descriptor: ServiceDescriptorProto, source: TypescriptFile): ts.ClassDeclaration {
         const
             interpreterType = this.interpreter.getServiceType(descriptor),
-            styleName = rpc.ClientStyle[this.style].toLowerCase(),
+            styleName = ClientStyle[this.style].toLowerCase().replace('_client', ''),
             ServiceType = this.imports.type(descriptor),
             ServiceClient = this.imports.type(descriptor, `${styleName}-client`),
             IServiceClient = this.imports.type(descriptor, `${styleName}-client-interface`),