+++ packages/test-generated/spec/service-example.spec.ts
@@ -1,9 +1,4 @@
-import {
-    ExampleRequest,
-    ExampleResponse,
-    ExampleServiceClient,
-    IExampleServiceClient
-} from "../ts-out/service-example";
+import {ExampleRequest, ExampleResponse,} from "../ts-out/service-example";
 import {
     ClientStreamingCall,
     Deferred,
@@ -19,6 +14,7 @@ import {
     ServerStreamingCall,
     UnaryCall
 } from "@protobuf-ts/runtime-rpc";
+import {ExampleServiceClient, IExampleServiceClient} from "../ts-out/service-example.client";
 
 
 describe('ExampleServiceClient', function () {